Remove runtime environment lookup
In theory, looking up the environment variables set on the service added an additional layer of security. In practice, it significantly hinders adoption and worsens it. Users are giving their service accounts project owner or editor because they don't know what kinds of permissions to grant. This drops all runtime environment configuration which greatly simplifies the codebase and improves the UX.